Sepet devam ediyor

+++ PlakDukkani.ViewModel/CartViewModels/Cart.cs
@@ -36,6 +36,6 @@ namespace PlakDukkani.ViewModel.CartViewModels
             }
         }
 
-        public decimal TotalQuantity => sepet.Values.Sum(a => a.Quantity);
+        public int TotalQuantity => sepet.Values.Sum(a => a.Quantity);
     }
 }
